moved the code around, made some components
